Move Benchmark module to Cryptol.Utils
